requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
prisma==0.10.0
supabase==1.0.3
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.0
python-multipart==0.0.6
httpx==0.24.1